Add teaching authority contactable
This represents something that holds a teaching authority which can be
contacted. For now this is only regions, but we're going to be extending
this to countries.
