+++ app/views/shared/_teaching_authority_contactable.html.erb
@@ -1,0 +1,19 @@
+<% if contactable.teaching_authority_address.present? %>
+  <p class="govuk-body app-teaching-authority-address" lang="<%= contactable.country.code %>"><%= contactable.teaching_authority_address %></p>
+<% end %>
+
+<% if contactable.teaching_authority_emails.present? %>
+  <p class="govuk-body">
+    <% contactable.teaching_authority_emails.each do |email| %>
+      <%= govuk_link_to email, "mailto:#{email}" %><br />
+    <% end %>
+  </p>
+<% end %>
+
+<% if contactable.teaching_authority_websites.present? %>
+  <p class="govuk-body">
+    <% contactable.teaching_authority_websites.each do |website| %>
+      <%= govuk_link_to website, website %><br />
+    <% end %>
+  </p>
+<% end %>
