+++ app/views/shared/_teaching_authority_contactable_fields.html.erb
@@ -1,0 +1,5 @@
+<%= f.govuk_text_area :teaching_authority_address, label: { text: "Teaching authority address" } %>
+
+<%= f.govuk_text_area :teaching_authority_emails_string, label: { text: "Teaching authority emails" }, hint: { text: "One on each line." } %>
+
+<%= f.govuk_text_area :teaching_authority_websites_string, label: { text: "Teaching authority websites" }, hint: { text: "One on each line." } %>
